fix: redshift-kompatibilität — double precision + metricflow time spine
mask_pii: DOUBLE → DOUBLE PRECISION (Redshift kennt keinen DOUBLE-Alias).
metricflow_time_spine: Redshift-Branch via Cross-Join-Digits statt generate_series.

+++ dbt/macros/mask_pii.sql
@@ -4,7 +4,7 @@
     {%- elif method == 'truncate_zip' -%}
         LEFT(CAST({{ column_name }} AS VARCHAR), 3)
     {%- elif method == 'round_coords' -%}
-        ROUND(CAST({{ column_name }} AS DOUBLE), 2)
+        ROUND(CAST({{ column_name }} AS DOUBLE PRECISION), 2)
     {%- else -%}
         {{ exceptions.raise_compiler_error("mask_pii: unknown method '" ~ method ~ "'. Use hash, truncate_zip, or round_coords.") }}
     {%- endif %}
